requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==21.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
